// src/types/partials/ImmunizationRecommendationRecommendation.ts
/* eslint-disable no-unused-vars */
/* eslint-disable no-trailing-spaces */
// This file is auto-generated by generate_types so do not edit manually

import { TExtension } from '../partials/Extension';
import { TCodeableConcept } from '../partials/CodeableConcept';
import { TImmunizationRecommendationDateCriterion } from '../partials/ImmunizationRecommendationDateCriterion';
import { TInt } from '../simpleTypes/Int';
import { TReference } from '../partials/Reference';

export type TImmunizationRecommendationRecommendation = {
    id?: String;
    extension?: TExtension[];
    modifierExtension?: TExtension[];
    vaccineCode?: TCodeableConcept[];
    targetDisease?: TCodeableConcept;
    contraindicatedVaccineCode?: TCodeableConcept[];
    forecastStatus: TCodeableConcept;
    forecastReason?: TCodeableConcept[];
    dateCriterion?: TImmunizationRecommendationDateCriterion[];
    description?: String;
    series?: String;
    doseNumberPositiveInt?: TInt;
    doseNumberString?: String;
    seriesDosesPositiveInt?: TInt;
    seriesDosesString?: String;
    supportingImmunization?: TReference[];
    supportingPatientInformation?: TReference[];
};